Rework NodeTicket constructor for NodeAddr inputs
It makes more sense to me to have the `NodeAddr` based constructor be
the default one and provide the string variant as static `parse` method
on the object. Most people will probably use the `NodeAddr` variant,
unless they already have something to parse, in which case the docs will
point them at the `parse` method.

Signed-off-by: Niklas Wimmer <mail@nwimmer.me>